production/docker/requirements.txt
pandas
numpy
scikit-learn==1.2.1
joblib
fastapi
uvicorn[standard]
google-generativeai